requirements.txt
absl-py==2.1.0
accelerate==0.28.0
acd==0.0.2
aiohttp==3.9.3
aiosignal==1.3.1
aix360==0.3.0
anchor-exp==0.0.2.0
annotated-types==0.6.0
anyio==4.2.0
anytree==2.12.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bleach==6.1.0
blis==0.7.11
boto3==1.34.71
botocore==1.34.71
bpemb==0.3.5
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chefboost==0.0.17
clarabel==0.7.1
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
confection==0.1.4
conllu==4.5.3
contourpy==1.2.0
cvxpy==1.4.2
cycler==0.12.1
cymem==2.0.8
datasets==2.16.1
debugpy==1.8.0
decorator==5.1.1
deeplift==0.6.13.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
dm-tree==0.1.8
docopt==0.6.2
ecos==2.0.13
editdistance==0.8.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flair==0.13.1
flatbuffers==23.5.26
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.0
gast==0.5.4
gdown==5.1.0
gensim==4.3.2
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googletrans==3.0.0
grpcio==1.60.1
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.20.3
hyperframe==5.2.0
idna==2.10
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
install==1.3.5
ipykernel==6.29.2
ipython==8.18.1
ipywidgets==8.1.2
isoduration==20.11.0
Janome==0.5.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
keras==2.15.0
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
language-tool-python==2.7.3
lazy_loader==0.3
lemminflect==0.2.3
libclang==16.0.6
lime==0.2.0.1
llvmlite==0.42.0
lru-dict==1.3.0
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==10.2.0
mpld3==0.5.10
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nlpaug==1.1.11
nltk==3.8.1
notebook==7.1.0
notebook_shim==0.2.3
num2words==0.5.13
numba==0.59.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
OpenHowNet==2.0
openpyxl==3.1.2
opt-einsum==3.3.0
osqp==0.6.5
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pillow==10.2.0
pinyin==0.4.0
platformdirs==4.2.0
pptree==3.1
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch_revgrad==0.2.0
pytz==2024.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qdldl==0.1.7.post0
qii-tool==0.1.3
qtconsole==5.5.1
QtPy==2.4.1
quantus==0.5.3
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
sage-importance==0.0.6
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.11.4
scs==3.2.4.post1
seaborn==0.13.2
segtok==1.5.11
semver==3.0.2
Send2Trash==1.8.2
sentencepiece==0.1.99
shap==0.44.1
six==1.16.0
slicer==0.0.7
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlitedict==2.1.0
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.23.0
termcolor==2.4.0
terminado==0.18.0
terminaltables==3.1.10
textattack==0.3.10
textaugment==2.0.0
textblob==0.18.0.post0
tf-explain==0.3.1
thinc==8.2.3
threadpoolctl==3.3.0
tifffile==2024.2.12
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformer-smaller-training-vocab==0.3.3
transformers==4.39.1
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==1.26.18
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
Wikipedia-API==0.6.0
word2number==1.1
wrapt==1.14.1
xgboost==2.0.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
